Fixes issue #552 (7000 threads in 15 minutes during video calls). The solution keeps a native thread in JNICaptureObserver alive with a Java thread attached to it and uses it to deliver the captured Image instead of creating a new Java thread for each delivery.
Unfortunately, the JNICaptureObserver instances leak - both in the old and the new implementation - and thus the new implementation leaks the mentioned thread being kept alive (and its attached Java thread). The leaking of the JNICaptureObserver instances will be fixed in a subsequent revision (and it will automatically fix the leaked threads) because it's a separate issue (as previously said, affecting the old implementation as well).