Implement a new cache handling for ZRTP
The new ZRTP cache handling implements two main topics.

The first modification reduces the number of write requests to the external
storage and simplifies overall handling of a ZRTP cache entry in the stack.

The second modification implements a new handling of the RS2 data. This will
be a topic in an upcoming revised ZRTP specification.