Add contact dialog: rename name and address labels
For clarity the labels for the display name and the contact's ID or
number have been renamed.

'Display name' becomes just 'Name', because 'Display Name' is
unnecessary technical terminology, and since there is no other Name
fields in the dialog, the technical distinction is unnecessary.

'ID or number' becomes 'IM address or phone number'. Refering to a phone
number is much clearer than to (any) number. The best terminology for
"ID" is less clear. However, 'IM address' is similar to 'email address'
and therefore to something users might know.

Thanks to David Bolton for suggesting this change. This commit message
contains parts of his rational for it.