Implementation of getServerChatRoomList.
Implemented getServerChatRoomList. There is no support for listing all
channels in the API, so a custom implementation is used with raw IRC
commands.
An attempt has been made to make this thread-safe, since the 'List'
button can be called multiple times, and whenever the search box
contents change, the original list is rerequested.