Initial support for secure connections and Wizard UI tweaks.
Initial support for secure connections has been introduced. An extra
package path has been added to the manifest in order to acquire
SSL-related classes. The IrcStack has been modified to receive an extra
parameter indicating whether or not to establish a secure connection.

A number of small UI tweaks have been made to the wizard. User id and
server entered into the "simple form" are now migrated to the full form.
An extra checkbox is added to indicate plaintext/secure connection. The
wizard currently defaults to secure. An appropriate default port is set
based on plaintext/secure type of connection. Registered new listeners
in order to evalute access to Next-button on every change.