Add contact dialog: reorganize input fields
To let a user enter what is most import to her or him, the dialog now
starts with the name for a new contact. Almost every contact manager
does this (take a look at email, address books (even paper address
books), IM clients, etc.) This way the user starts with what they know.

Second is the ID field which keeps the actual contact information.
Usually, a user will have a name and the address of new contact. Since
the name was entered first, (s)he will understand what has to be typed
into this field.

Third is the user's account. The group setting is last, because it is
optional.

Thanks to David Bolton for suggesting this change. This commit message
contains parts of his rational for it.