New implementation of chat room member role management.
New implementation of chat room member role management that determines
active role based on set channel-user-modes. Multiple modes can be
active, such as Voice and Operator. Active role is determined based on
the strongest channel-user-mode.

In case the strongest mode is modified, a local user role change event
is fired. In case a weaker mode is modified, a system message is shown
to update the user on the mode change, but the role does not change.