Fix for inserting large number of query results for server chat room listing.
Previously chat rooms were inserted one by one into the result and
listeners were informed for each of the chat rooms immediately after it
was inserted, and on top of that we reorder after each insertion. All in
all this caused very unpredictable results.

The modification massively inserts the whole result set and then fires a
listener event for each of the added chat rooms.