Changed BasicInstantMessagingTransport to extend OperationSet directly.
Based on my own doubts and Ingo's feedback I've decided to not base
OperationSetBasicInstantMessagingTransport on
OperationSetBasicInstantMessaging. Even though it doesn't make much
sense to implement BasicInstantMessagingTransport and not
BasicInstantMessaging. In theory we could at least query the
characteristics if not actually use the protocol, so I've decided to
extend OperationSet directly.