Unknown commands now result in failed message delivery.
To, for one, protect users from accidents and, for two, don't
accidentally send "stupid messages" we do no longer send any messages
starting with '/'. Instead it shows a message delivery failed message
for error UNSUPPORTED_OPERATION. Reasoning is that the intended
operation is either unsupported or unknown. In both cases there is no
support for it, so let the user know.