Primitive implementation of presence watcher for Contacts.
A primitive initial implementation for periodically testing the presence
of contacts by using IRC's ISON query. Currently using a period of 60
seconds, it will check with the IRC server whether or not contacts are
online. Contact presence change events will be fired for each of the
results, both online and offline. For presence changes, the contact list
will be updated accordingly.

Additionally, presence update events will also be fired in case an
ERR_NOSUCHNICK is received. This indicates that something assumed that
the contact was online, but it turns out that this contact is actually
offline.