Improved error handling for commands.
In case of bad command usage, throw an IllegalArgumentException.
IllegalArgumentException will be caught and handled specially. Whenever
IAE is thrown, it will be caught and command.help() will be called so
that we can receive additional usage instructions. The error and the
help information will be thrown inside a BadCommandInvocationException
which will be handled by the appropriate OperationSet.

In case of such an event, a system message will be fired with the usage
instructions as to inform the user on how to use the command.  Other
exceptions will be logged as an error, as they are not expected and
considered to be an implementation issue.