Updated IRC client library irc-api to custom version to test bug fix.
There's this bug in irc-api where *every* channel message arrives as a
private message to the local user. This quickly results in lots of
notifications, a contact for every message sender and lots of new tabs
in the chat window. (See
http://lists.jitsi.org/pipermail/dev/2014-November/022755.html)

I believe that this is fixed now. The fix that is applied can be found
at https://code.google.com/p/irc-api/issues/detail?id=20 . While I am
waiting for upstream to incorporate the fix, I'll use this version and
ask George to test this out and see if the problem disappears.