Mostly complete implementation of IRC v3 capability negotiation and SASL authentication.
This commit implements IRC v3 capability negotiation. A client
configuration option is provided to explicitly enable/disable IRC v3.
IRCv3 support will be enabled by default and may be enabled in specific
cases when IRC v3 support is explicitly not allowed. In case IRC v3
support is enabled, we will in all cases trigger capability negotiation,
though at this time, if we do not want to negotiate for SASL
authentication, there is no extension to negotiate for. In that case we
will issue CAP END to ensure that IRC server knows of our IRC v3
capabilities but do not engage in actual capability negotiation. If we
do not issue any command, there is a chance that an IRC server will
downgrade interoperation capabilities to the original RFC1459 feature
set.