Enable GTKLookAndFeel on Linux/FreeBSD globally
Right now, the native GTKLookAndFeel is only used when on a KDE session, which curiously enough is actually one of the few non-GTK-based DEs. The proposed changes enables GTKLookAndFeel to be used globally on Linux and FreeBSD, if the style is available, and gives users a much higher integration regardless their desktop environment of choice.

If there is a compelling reason on why not to enable it for other environments please tell me so I can look into another possible solution, because the out-of-the-box Jitsi UI experience on Linux could be much smoother. Thank you!