Implementation of using Basic Poller as the fall back scenario when the MONITOR registration limit is reached.
This implementation enables both the Monitor presence watcher and the
Basic Poller presence watcher. Basic Poller serves as a fall back
mechanism when nicks cannot be added to the MONITOR reg list. Basic
Poller runs in the background and functions on the difference of the
set containing all nicks and the set containing all nicks registered
with MONITOR. Typically, Basic Poller will only ever do anything useful
after a MONITOR registration failed. Either because of a full list or
some other error.