Fixes an issue with the toop tip of contacts having large avatars.
When one hovers over a contact from his contact list, a pop up tool tip
appears. The tool tip used to show the full avatar image instead of a
scaled one. The problem with that approach was that:
1] The large image might have taken a huge amount of space in the
display area
2] The large image takes a lot of time to be drawn on the screen and
this block the whole UI.

This commit fixes the aforementioned issue. The avatar that is shown in
the tool tip is now scaled and cached accordingly.