Add PUBLIC_NAT_IP
A public IP that should be distributed to peers as a contact address
instead of the locally determined one. This is required in the case
of a 1:1 NAT, e.g. within AWS EC2, where the local interfaces do not
correspond with the externally visible address.

Applies to ICE candidates as well as SIP URIs.