Updates debian dependencies.
java.lang.UnsatisfiedLinkError: no splashscreen in java.library.path
	at net.java.sip.communicator.impl.splashscreen.SplashScreenActivator.start(SplashScreenActivator.java:70)
is printed otherwise. However, because Jitsi Desktop starts even with this error, the splash screen is optional. Consequently, a headless Java environment is sufficient and the non-headless Java environment is optional. Therefore, the non-headless is not in Depends but in Recommends.